+++ util.js
@@ -13,7 +13,7 @@ module.exports = {
         if (!fs.existsSync(pth)) {
             fs.mkdirSync(pth, { recursive: true });
         }
-    
+
         fs.writeFileSync(fname, content.toString());
     },
 
@@ -25,10 +25,10 @@ module.exports = {
     getFileContent: function(fn) {
 
         if (!fs.existsSync(fn)) {
-            consoleError('File: `' + fn + '` not found.');
+            this.consoleError('File: `' + fn + '` not found.');
             process.exit(1);
         }
-    
+
         return fs.readFileSync(fn).toString();
 
     },
@@ -47,11 +47,11 @@ module.exports = {
     consoleSuccess: function(msg) {
         this.consoleMsg('\x1b[42m', '\x1b[37m', msg);
     },
-    
+
     consoleError: function(msg) {
         this.consoleMsg('\x1b[41m', '\x1b[37m', msg);
     },
-    
+
     consoleMsg: function(bg, fg, msg) {
         console.log(bg, fg, msg, '\x1b[0m');
         console.log('\n');